# gameyapi/openapi.yaml
openapi: 3.0.0
info:
  title: Game Y API
  version: 1.0.0
  description: >
    REST API for managing Game Y sessions—a triangular board game played on a
    hexagonal grid. The service manages game state, performs win detection using
    Union-Find logic, persists results to MongoDB, and interfaces with a Rust-based 
    AI engine for bot moves.

servers:
  - url: http://localhost:3001
    description: Local development server

tags:
  - name: Game
    description: Game session lifecycle and move management
  - name: User
    description: Player profiles and match history
  - name: Health
    description: Service and engine status

paths:
  /play/create:
    post:
      tags: [Game]
      summary: Create a new game session
      description: >
        Initializes a new game. If a JWT is provided in the Authorization header, 
        the game will be linked to that user ID for profile tracking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [hvh, hvb]
                  description: "hvh: Human vs Human, hvb: Human vs Bot."
                boardSize:
                  type: integer
                  minimum: 5
                  maximum: 15
                  default: 11
                difficulty:
                  type: string
                  enum: [beginner, medium, advanced]
                  default: medium
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'

  /play/{gameId}:
    get:
      tags: [Game]
      summary: Get game state
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current session data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'
    delete:
      tags: [Game]
      summary: Delete a session
      responses:
        '200':
          description: Game removed from memory

  /play/{gameId}/move:
    post:
      tags: [Game]
      summary: Submit a move
      description: >
        Submits a move for the current player. In 'hvb' mode, the bot move 
        is automatically triggered and returned in the response.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [player, x, y]
              properties:
                player:
                  type: integer
                  enum: [0, 1]
                x:
                  type: integer
                y:
                  type: integer
      responses:
        '200':
          description: Move processed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GameSession'
                  - type: object
                    properties:
                      botMove:
                        $ref: '#/components/schemas/Coords'

  /play/{gameId}/undo:
    post:
      tags: [Game]
      summary: Undo last move(s)
      description: Reverts 1 move in hvh, or 2 moves (bot + human) in hvb.
      responses:
        '200':
          description: State reverted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'

  /profile:
    get:
      tags: [User]
      summary: Get player statistics
      description: Returns win rate, best score, and history from MongoDB.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistics retrieved
        '401':
          description: Unauthorized / Missing JWT

  /health:
    get:
      tags: [Health]
      summary: Check service status
      responses:
        '200':
          description: Status of API, Rust Engine, and active session count

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    GameSession:
      type: object
      properties:
        gameId:
          type: string
        mode:
          type: string
        status:
          type: string
          enum: [ongoing, finished]
        currentPlayer:
          type: integer
        winner:
          type: integer
          nullable: true
        winningPath:
          type: array
          items:
            $ref: '#/components/schemas/Coords'
        layout:
          type: string
          description: "YEN format layout string"

    Coords:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer

    Error:
      type: object
      properties:
        error:
          type: string